Move HeaderFooter to Traits
