+++ src/Views/Traits/Helpers/ColumnHelpers.php
@@ -319,104 +319,7 @@ trait ColumnHelpers
         return $this->selected === true;
     }
 
-    public function hasSecondaryHeader(): bool
-    {
-        return $this->secondaryHeader === true;
-    }
 
-    public function hasSecondaryHeaderCallback(): bool
-    {
-        return $this->secondaryHeaderCallback !== null;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getSecondaryHeaderCallback()
-    {
-        return $this->secondaryHeaderCallback;
-    }
-
-    /**
-     * @param  mixed  $rows
-     * @return mixed
-     */
-    public function getSecondaryHeaderContents($rows, array $filterGenericData)
-    {
-        $value = null;
-        $callback = $this->getSecondaryHeaderCallback();
-
-        if ($this->hasSecondaryHeaderCallback()) {
-            if (is_callable($callback)) {
-                $value = call_user_func($callback, $rows);
-                if ($this->isHtml()) {
-                    return new HtmlString($value);
-                }
-            } elseif ($callback instanceof Filter) {
-                return $callback->setFilterPosition('header')->setGenericDisplayData($filterGenericData)->render();
-            } elseif (is_string($callback)) {
-                $filter = $this->getComponent()->getFilterByKey($callback);
-
-                if ($filter instanceof Filter) {
-                    return $filter->setFilterPosition('header')->setGenericDisplayData($filterGenericData)->render();
-                }
-            } else {
-                throw new DataTableConfigurationException('The secondary header callback must be a closure, filter object, or filter key if using secondaryHeaderFilter().');
-            }
-        }
-
-        return $value;
-    }
-
-    public function hasFooter(): bool
-    {
-        return $this->footer === true;
-    }
-
-    public function hasFooterCallback(): bool
-    {
-        return $this->footerCallback !== null;
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getFooterCallback()
-    {
-        return $this->footerCallback;
-    }
-
-    /**
-     * @param  mixed  $rows
-     * @return mixed
-     */
-    public function getFooterContents($rows, array $filterGenericData)
-    {
-        $value = null;
-        $callback = $this->getFooterCallback();
-
-        if ($this->hasFooterCallback()) {
-            if (is_callable($callback)) {
-                $value = call_user_func($callback, $rows);
-
-                if ($this->isHtml()) {
-                    return new HtmlString($value);
-                }
-            } elseif ($callback instanceof Filter) {
-                return $callback->setFilterPosition('footer')->setGenericDisplayData($filterGenericData)->render();
-            } elseif (is_string($callback)) {
-                $filter = $this->getComponent()->getFilterByKey($callback);
-
-                if ($filter instanceof Filter) {
-                    return $filter->setFilterPosition('footer')->setGenericDisplayData($filterGenericData)->render();
-                }
-            } else {
-                throw new DataTableConfigurationException('The footer callback must be a closure, filter object, or filter key if using footerFilter().');
-            }
-        }
-
-        return $value;
-    }
 
     public function isClickable(): bool
     {
